removing need for editing yml files when updating deploy
run_compose command now has this extra command for `deploy-localhost`
that will create (ask to override if existing) a
docker-compose.override.yml file containing the localhost port mapping
for the services listed in the services_in_use directory.

So now the yml files no longer need the localhost port mapping (these
should be changed in the script if neeeded, or if new services are
added).

For our deployment process this would mean that after updating a service
yml file we can just replace (or symlink) the files in services_in_use
with the new version, and that's it.

To test on a localhost setup it now requires the user to use first the
deploy-localhost command, which will setup the override file with the
necessary port mappings.
refs #38